feat: add container health check to LB service
