+++ internal/pkg/manifest/workload.go
@@ -10,12 +10,14 @@ import (
 	"path/filepath"
 	"reflect"
 	"strconv"
+	"time"
 
 	"github.com/imdario/mergo"
 
 	"github.com/google/shlex"
 
 	"github.com/aws/aws-sdk-go/aws"
+	"github.com/aws/aws-sdk-go/service/ecs"
 	"gopkg.in/yaml.v3"
 )
 
@@ -110,6 +112,12 @@ func transformImage() func(dst, src reflect.Value) error {
 		}
 		return nil
 	}
+}
+
+// ImageWithPortAndHealthcheck represents a container image with an exposed port and health check.
+type ImageWithPortAndHealthcheck struct {
+	ImageWithPort `yaml:",inline"`
+	HealthCheck   *ContainerHealthCheck `yaml:"healthcheck"`
 }
 
 // ImageWithPort represents a container image with an exposed port.
@@ -510,10 +518,6 @@ func UnmarshalWorkload(in []byte) (WorkloadManifest, error) {
 		if err := yaml.Unmarshal(in, m); err != nil {
 			return nil, fmt.Errorf("unmarshal to backend service: %w", err)
 		}
-		if m.BackendServiceConfig.ImageConfig.HealthCheck != nil {
-			// Make sure that unset fields in the healthcheck gets a default value.
-			m.BackendServiceConfig.ImageConfig.HealthCheck.applyIfNotSet(newDefaultContainerHealthCheck())
-		}
 		return m, nil
 	case ScheduledJobType:
 		m := newDefaultScheduledJob()
@@ -523,6 +527,82 @@ func UnmarshalWorkload(in []byte) (WorkloadManifest, error) {
 		return m, nil
 	default:
 		return nil, &ErrInvalidWorkloadType{Type: typeVal}
+	}
+}
+
+// ContainerHealthCheck holds the configuration to determine if the service container is healthy.
+// See https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-ecs-taskdefinition-healthcheck.html
+type ContainerHealthCheck struct {
+	Command     []string       `yaml:"command"`
+	Interval    *time.Duration `yaml:"interval"`
+	Retries     *int           `yaml:"retries"`
+	Timeout     *time.Duration `yaml:"timeout"`
+	StartPeriod *time.Duration `yaml:"start_period"`
+}
+
+// newDefaultContainerHealthCheck returns container health check configuration
+// that's identical to a load balanced web service's defaults.
+func newDefaultContainerHealthCheck() *ContainerHealthCheck {
+	return &ContainerHealthCheck{
+		Command:     []string{"CMD-SHELL", "curl -f http://localhost/ || exit 1"},
+		Interval:    durationp(10 * time.Second),
+		Retries:     aws.Int(2),
+		Timeout:     durationp(5 * time.Second),
+		StartPeriod: durationp(0 * time.Second),
+	}
+}
+
+// apply overrides the healthcheck's fields if other has them set.
+func (hc *ContainerHealthCheck) apply(other *ContainerHealthCheck) {
+	if other.Command != nil {
+		hc.Command = other.Command
+	}
+	if other.Interval != nil {
+		hc.Interval = other.Interval
+	}
+	if other.Retries != nil {
+		hc.Retries = other.Retries
+	}
+	if other.Timeout != nil {
+		hc.Timeout = other.Timeout
+	}
+	if other.StartPeriod != nil {
+		hc.StartPeriod = other.StartPeriod
+	}
+}
+
+// applyIfNotSet changes the healthcheck's fields only if they were not set and the other healthcheck has them set.
+func (hc *ContainerHealthCheck) applyIfNotSet(other *ContainerHealthCheck) {
+	if hc.Command == nil && other.Command != nil {
+		hc.Command = other.Command
+	}
+	if hc.Interval == nil && other.Interval != nil {
+		hc.Interval = other.Interval
+	}
+	if hc.Retries == nil && other.Retries != nil {
+		hc.Retries = other.Retries
+	}
+	if hc.Timeout == nil && other.Timeout != nil {
+		hc.Timeout = other.Timeout
+	}
+	if hc.StartPeriod == nil && other.StartPeriod != nil {
+		hc.StartPeriod = other.StartPeriod
+	}
+}
+
+// HealthCheckOpts converts the image's healthcheck configuration into a format parsable by the templates pkg.
+func (i ImageWithPortAndHealthcheck) HealthCheckOpts() *ecs.HealthCheck {
+	if i.HealthCheck == nil {
+		return nil
+	}
+	// Make sure that unset fields in the healthcheck gets a default value.
+	i.HealthCheck.applyIfNotSet(newDefaultContainerHealthCheck())
+	return &ecs.HealthCheck{
+		Command:     aws.StringSlice(i.HealthCheck.Command),
+		Interval:    aws.Int64(int64(i.HealthCheck.Interval.Seconds())),
+		Retries:     aws.Int64(int64(*i.HealthCheck.Retries)),
+		StartPeriod: aws.Int64(int64(i.HealthCheck.StartPeriod.Seconds())),
+		Timeout:     aws.Int64(int64(i.HealthCheck.Timeout.Seconds())),
 	}
 }
